Use uint8_t for flag d1_pig_present

+++ common/main/gamemine.h
@@ -143,9 +143,11 @@ int load_mine_data_compiled(NamedPHYSFS_File LoadFile, const char *Gamesave_curr
 
 #if defined(DXX_BUILD_DESCENT_II)
 namespace dsx {
-extern int d1_pig_present;
+
+extern uint8_t d1_pig_present;
 
 /* stuff for loading descent.pig of descent 1 */
 uint16_t convert_d1_tmap_num(uint16_t d1_tmap_num);
+
 }
 #endif
